First (big!) commit
This app will (hopefully!) be a cross platform streaming app for video streaming sites like youtube and anime/movie streaming sites like bflix.gg or hianime.to. The name "VidAiO" = Video All In One is meant to indicate its ability to play video from any site, on any platform.

For now, however, this is just a very basic linux flatpak app to search "hianime.to", built with QT and QML. The results and episode list are served in a local website, served by golang Echo, and displayed in QtWebView. In golang I tried to implement seperation of concerns between the API handler and the UI, as the APIFetcher struct and the structs implementing the "HTMLRetriever" to get the HTML are seperate. However, for now, all we really do is just get some JSON from hianime.to, and display the "html" property of JSON as raw html, with some CSS styling to make it actually presentable. To make the html templates served by golang cleaner and less repetetive, I seperated the head and end of the body, which I placed in a seperate, "base.html" file.